requirements.txt
requests~=2.24.0
PyQt5~=5.15.1
getmac~=0.8.2
python-magic~=0.4.18
setuptools~=50.3.0